wireless-regdb: Add 160 MHz channel bandwidth for Dominican Republic (DO)
This extends regulatory rules for Dominican Republic to allow 160 MHz
VHT channels to be used in 5.15-5.25 GHz, 5.25-5.35 GHz. Since the rules
between 5.15-5.25 and 5.25-5.35 GHz ranges are different (e.g., DFS
requirement), the AUTO-BW flag needs to be used to allow the maximum
channel bandwidth to be determined automatically based on two rules.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>