wireless-regdb: Add 160 MHz channel bandwidth for Algeria (DZ)
This extends regulatory rules for Algeria to allow 160 MHz VHT channels
to be used in 5.15-5.25 GHz, 5.25-5.35 GHz, and 5.47-5.725 GHz ranges.
Since the rules between 5.15-5.25 and 5.25-5.35 GHz ranges are different
in most cases (e.g., DFS requirement), the AUTO-BW flag needs to be used
to allow the maximum channel bandwidth to be determined automatically
based on two rules. The larger 5.47-5.725 GHz range is large enough to
fit the full 160 MHz channel in a single rule, so that does not require
similar handling and can just change the maximum allowed bandwidth to
160 MHz in the rule.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>
Signed-off-by: John W. Linville <linville@tuxdriver.com>