wireless-regdb: Update U-NII-2c (5470 ~ 5725 MHz) rules for Taiwan (TW)
Taiwan's Ministry of Transportation and Communications revised its
frequency allocation rules [1] on 2014/11/17, allowing usage of 5600 ~
5650 MHz, previously allocated to weather radars, to U-NII applications
with DFS support.

Also, the technical regulations [2] show that for 5470 ~ 5725 MHz U-NII
applications, the peak transmit power shall not exceed the lesser of
250 mW (slightly less than 24 dBm) or 11 dBm + 10log B, where B is the
26dB emission bandwidth in MHz. This is slightly more than 23 dBm for
20 MHz channels.

This patch updates both. Also add links to the two documents into the
database.

[1] http://www.motc.gov.tw/websitedowndoc?file=post/201411171137330.doc&filedisplay=Table+of+radio+frequency+allocation.doc
[2] http://www.ncc.gov.tw/english/show_file.aspx?table_name=news&file_sn=681

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>