wireless-regdb: Update 5GHz rules for US
The FCC increased the maximum conducted transmission power for the
U-NII-1 (5150 ~ 5250 MHz) band to 30 dBm or 1 W for master devices
and 24 dBm or 250 mW for mobile/portable
devices.

Since 250 mW is slightly lower than 24 dBm, set the power limit to
a conservative value of 23 dBm.

Effective 6/2/2014.

See FCC KDB 905462 D06.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>