wireless-regdb: Update regulatory rules for Germany (DE) on 5GHz
In Europe ETSI standardized the used for short range devices
(SRD) [1] in ETSI EN 300 440-1 [2]. According to this standard
generic use equipment is allowed to transmit in the frequency
range form 5725 MHz to 5875 MHz with a maximum output power of 25
mW e.i.r.p. This generic allocation also allows transmission of
802.11 devices.

[3] has a list of countries and their status about the
implementation of the SRD frequency bands. For the 5 GHz band
this can be seen on page 38 in the paragraph ANNEX 1 and the
entry Annex j.

Attached is a patch that adds the 5 GHz SRD band to db.txt for
Germany.

[1] http://www.etsi.org/technologies-clusters/technologies/radio/short-range-devices
[2] http://www.etsi.org/deliver/etsi_en/300400_300499/30044001/01.06.01_60/en_30044001v010601p.pdf
[3] http://www.erodocdb.dk/Docs/doc98/official/pdf/REC7003e.pdf

Signed-off-by: Maximilian Engelhardt <engelhardt@perisens.de>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>