wireless-regdb: Update regulatory references for Germany (DE)
The "Frequenznutzungsplan" is not really a good reference as it's
a list of all frequency allocations and does only list basic
information about the allocation. This patch changes the comments
to reference directly the corresponding documents which provide
all information about the allocation.

This patch does only change comments, all entries have been left
unchanged. However the entries have been verified to match the
mentioned documents.

Signed-off-by: Maximilian Engelhardt <engelhardt@perisens.de>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>