wireless-regdb: Update rules for Turkey 5ghz
Based on the information from [1] the power limit for 5170-5250 Mhz
can be increased. Additionally 5170-5330Mhz are restricted to indoor
use.

There are references to e.i.r.p density in [1] that will affect the
power available for wide channels, but they are not reflected in
this change as it's unclear how they would be specified.

[1] https://www.btk.gov.tr/File/?path=ROOT%2f1%2fDocuments%2fOrdinance%2fBY%2DLAW%20ON%20SHORT%20RANGE%20DEVICES.pdf

Signed-off-by: Edwin Steele <edwin@wordspeak.org>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>